+++ src/main/java/com/forgestove/create_cyber_goggles/CreateCyberGoggles.java
@@ -1,17 +1,42 @@
 package com.forgestove.create_cyber_goggles;
-import com.forgestove.create_cyber_goggles.content.ModConfig;
-import me.shedaniel.autoconfig.AutoConfig;
-import me.shedaniel.autoconfig.serializer.Toml4jConfigSerializer;
-import net.minecraftforge.client.ConfigScreenHandler.ConfigScreenFactory;
-import net.minecraftforge.fml.ModLoadingContext;
+import com.forgestove.create_cyber_goggles.content.config.CyberConfig;
+import com.forgestove.create_cyber_goggles.content.event.*;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.client.event.InputEvent.*;
+import net.minecraftforge.client.event.*;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
+import net.minecraftforge.fml.common.Mod.EventBusSubscriber.Bus;
 @Mod(CreateCyberGoggles.ID)
 public class CreateCyberGoggles {
 	public static final String ID = "create_cyber_goggles";
-	public static ModConfig config;
 	public CreateCyberGoggles() {
-		config = AutoConfig.register(ModConfig.class, Toml4jConfigSerializer::new).getConfig();
-		var factory = new ConfigScreenFactory((mc, screen) -> AutoConfig.getConfigScreen(ModConfig.class, screen).get());
-		ModLoadingContext.get().registerExtensionPoint(ConfigScreenFactory.class, () -> factory);
+		CyberConfig.register();
+	}
+	@EventBusSubscriber(modid = ID, value = Dist.CLIENT, bus = Bus.FORGE)
+	public static class ClientGameEvents {
+		@SubscribeEvent
+		public static void key(Key event) {
+			KeyInput.toggleDiving();
+			KeyInput.openConfigScreen();
+			KeyInput.openStockScreen();
+			KeyInput.previewFilterScreen();
+		}
+		@SubscribeEvent
+		public static void mouseScrollingEvent(MouseScrollingEvent event) {
+			MouseScroll.onMouseScroll(event);
+		}
+	}
+	@EventBusSubscriber(modid = ID, value = Dist.CLIENT, bus = Bus.MOD)
+	public static class ClientModEvents {
+		@SubscribeEvent
+		public static void registerKeyMappingsEvent(RegisterKeyMappingsEvent event) {
+			ModKeyMapping.register(event);
+		}
+		@SubscribeEvent
+		public static void registerGuiLayersEvent(RegisterGuiOverlaysEvent event) {
+			OverlayRenderer.register(event);
+		}
 	}
 }