+++ src/main/java/com/forgestove/create_cyber_goggles/content/config/CyberConfigData.java
@@ -1,11 +1,11 @@
-package com.forgestove.create_cyber_goggles.content;
+package com.forgestove.create_cyber_goggles.content.config;
 import com.forgestove.create_cyber_goggles.CreateCyberGoggles;
 import me.shedaniel.autoconfig.ConfigData;
 import me.shedaniel.autoconfig.annotation.Config;
 import me.shedaniel.autoconfig.annotation.ConfigEntry.Category;
 import me.shedaniel.autoconfig.annotation.ConfigEntry.Gui.*;
 @Config(name = CreateCyberGoggles.ID)
-public class ModConfig implements ConfigData {
+public class CyberConfigData implements ConfigData {
 	@Category("armor") @TransitiveObject public Armor armor = new Armor();
 	@Category("chainConveyor") @TransitiveObject public ChainConveyor chainConveyor = new ChainConveyor();
 	@Category("flywheel") @TransitiveObject public Flywheel flywheel = new Flywheel();
@@ -30,15 +30,16 @@ public class ModConfig implements ConfigData {
 		@Tooltip public boolean forcedBackend = false;
 	}
 	public static class Goggles {
-		@Tooltip public boolean enableInSurvival = true;
-		@Tooltip public boolean enableInCreative = true;
-		@Tooltip public boolean enableInSpectator = true;
-		@Tooltip public boolean enableInAdventure = true;
 		@Tooltip public boolean enhancedInfo = true;
+		@Tooltip public boolean hideStaticKineticInfo = false;
 		@Tooltip public boolean enhancedStoreRender = true;
 		@Tooltip public boolean renderExtraItems = true;
 		@Tooltip public boolean enableKineticEffect = true;
 		@Tooltip public boolean preciseNumbers = true;
+		@Tooltip public boolean enableInSurvival = true;
+		@Tooltip public boolean enableInCreative = true;
+		@Tooltip public boolean enableInSpectator = true;
+		@Tooltip public boolean enableInAdventure = true;
 	}
 	public static class Jei {
 		@Tooltip @RequiresRestart public boolean nonrandomScrap = true;