+++ src/main/java/com/forgestove/create_cyber_goggles/content/event/KeyInput.java
@@ -1,6 +1,7 @@
 package com.forgestove.create_cyber_goggles.content.event;
-import com.forgestove.create_cyber_goggles.*;
-import com.forgestove.create_cyber_goggles.content.ModConfig;
+import com.forgestove.create_cyber_goggles.CreateCyberGoggles;
+import com.forgestove.create_cyber_goggles.content.config.*;
+import com.forgestove.create_cyber_goggles.content.util.RunSafely;
 import com.simibubi.create.AllMenuTypes;
 import com.simibubi.create.content.logistics.filter.*;
 import com.simibubi.create.content.logistics.stockTicker.*;
@@ -16,16 +17,23 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.HitResult.Type;
 
+import java.lang.invoke.*;
 import java.util.Collections;
+@SuppressWarnings("JavaLangInvokeHandleSignature")
 public class KeyInput {
+	public static final MethodHandle FILTER_TYPE = RunSafely.run(() -> MethodHandles.lookup().findGetter(
+		FilterItem.class,
+		"type",
+		Enum.class
+	));
 	public static StockTickerBlockEntity lastBlockEntity = null;
 	public static void toggleDiving() {
 		if (!ModKeyMapping.TOGGLE_DIVING.get().isDown()) return;
 		var mc = Minecraft.getInstance();
 		var player = mc.player;
 		if (player == null || mc.screen != null) return;
-		var enabled = CreateCyberGoggles.config.armor.removeDivingBootsAffect;
-		CreateCyberGoggles.config.armor.removeDivingBootsAffect = !enabled;
+		var enabled = CyberConfig.get().armor.removeDivingBootsAffect;
+		CyberConfig.get().armor.removeDivingBootsAffect = !enabled;
 		player.displayClientMessage(
 			Component.translatable("message.%s.%s".formatted(
 				CreateCyberGoggles.ID,
@@ -37,7 +45,7 @@ public class KeyInput {
 		if (!ModKeyMapping.OPEN_CONFIG.get().isDown()) return;
 		var mc = Minecraft.getInstance();
 		if (mc.screen != null) return;
-		mc.setScreen(AutoConfig.getConfigScreen(ModConfig.class, null).get());
+		mc.setScreen(AutoConfig.getConfigScreen(CyberConfigData.class, null).get());
 	}
 	public static void openStockScreen() {
 		if (!ModKeyMapping.OPEN_STOCK.get().isDown()) return;
@@ -81,27 +89,24 @@ public class KeyInput {
 		}
 	}
 	public static void setFilterScreen(ItemStack filter) {
-		try {
-			var field = FilterItem.class.getDeclaredField("type");
-			field.setAccessible(true);
-			if (!(filter.getItem() instanceof FilterItem filterItem)) return;
-			if (!field.getType().isEnum()) return;
-			var mc = Minecraft.getInstance();
-			var player = mc.player;
-			if (player == null) return;
-			var inv = player.getInventory();
-			var name = filter.getHoverName();
-			Screen screen;
-			switch (((Enum<?>) field.get(filterItem)).ordinal()) {
-				case 0 -> screen = new FilterScreen(FilterMenu.create(-1, inv, filter), inv, name);
-				case 1 -> screen = new AttributeFilterScreen(AttributeFilterMenu.create(-1, inv, filter), inv, name);
-				case 2 -> screen = new PackageFilterScreen(PackageFilterMenu.create(-1, inv, filter), inv, name);
-				default -> {
-					return;
-				}
+		if (!(filter.getItem() instanceof FilterItem filterItem)) return;
+		var mc = Minecraft.getInstance();
+		var player = mc.player;
+		if (player == null) return;
+		var inv = player.getInventory();
+		var name = filter.getHoverName();
+		if (FILTER_TYPE == null) return;
+		var type = RunSafely.run(() -> (Enum<?>) FILTER_TYPE.invokeExact(filterItem));
+		if (type == null) return;
+		Screen screen;
+		switch (type.ordinal()) {
+			case 0 -> screen = new FilterScreen(FilterMenu.create(-1, inv, filter), inv, name);
+			case 1 -> screen = new AttributeFilterScreen(AttributeFilterMenu.create(-1, inv, filter), inv, name);
+			case 2 -> screen = new PackageFilterScreen(PackageFilterMenu.create(-1, inv, filter), inv, name);
+			default -> {
+				return;
 			}
-			ScreenOpener.open(screen);
-		} catch (Exception ignored) {
 		}
+		ScreenOpener.open(screen);
 	}
 }