+++ src/main/java/com/forgestove/create_cyber_goggles/mixin/armor/DivingBootsItemMixin.java
@@ -1,5 +1,5 @@
 package com.forgestove.create_cyber_goggles.mixin.armor;
-import com.forgestove.create_cyber_goggles.CreateCyberGoggles;
+import com.forgestove.create_cyber_goggles.content.config.CyberConfig;
 import com.simibubi.create.content.equipment.armor.DivingBootsItem;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.*;
@@ -8,6 +8,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public abstract class DivingBootsItemMixin {
 	@Inject(method = "isWornBy", at = @At("HEAD"), remap = false, cancellable = true)
 	private static void isWornBy(CallbackInfoReturnable<Boolean> returnable) {
-		if (CreateCyberGoggles.config.armor.removeDivingBootsAffect) returnable.setReturnValue(false);
+		if (CyberConfig.get().armor.removeDivingBootsAffect) returnable.setReturnValue(false);
 	}
 }