+++ src/main/java/com/forgestove/create_cyber_goggles/mixin/chainConveyor/ChainConveyorRidingHandlerMixin.java
@@ -1,6 +1,6 @@
 package com.forgestove.create_cyber_goggles.mixin.chainConveyor;
-import com.forgestove.create_cyber_goggles.*;
-import com.forgestove.create_cyber_goggles.content.Util;
+import com.forgestove.create_cyber_goggles.content.config.CyberConfig;
+import com.forgestove.create_cyber_goggles.content.util.Common;
 import com.simibubi.create.AllPackets;
 import com.simibubi.create.AllTags.AllItemTags;
 import com.simibubi.create.content.kinetics.chainConveyor.*;
@@ -26,7 +26,7 @@ public abstract class ChainConveyorRidingHandlerMixin {
 		if (mc.isPaused()) return;
 		var player = mc.player;
 		if (player == null) return;
-		var chainConveyor = CreateCyberGoggles.config.chainConveyor;
+		var chainConveyor = CyberConfig.get().chainConveyor;
 		if (!chainConveyor.alwaysAllowRiding && !AllItemTags.CHAIN_RIDEABLE.matches(mc.player.getMainHandItem())) {
 			stopRiding();
 			return;
@@ -65,7 +65,7 @@ public abstract class ChainConveyorRidingHandlerMixin {
 			}
 		}
 		player.setDeltaMovement(player.getDeltaMovement().scale(0.75).add(diff.scale(0.25)));
-		if (Util.testForStealth(player)) player.connection.send(new ServerboundPlayerCommandPacket(player, Action.PRESS_SHIFT_KEY));
+		if (Common.testForStealth(player)) player.connection.send(new ServerboundPlayerCommandPacket(player, Action.PRESS_SHIFT_KEY));
 		if (AnimationTickHolder.getTicks() % 10 == 0)
 			AllPackets.getChannel().sendToServer(new ServerboundChainConveyorRidingPacket(ridingChainConveyor, false));
 	}