+++ src/main/java/com/forgestove/create_cyber_goggles/mixin/jei/SequencedAssemblyCategoryMixin.java
@@ -1,5 +1,5 @@
 package com.forgestove.create_cyber_goggles.mixin.jei;
-import com.forgestove.create_cyber_goggles.CreateCyberGoggles;
+import com.forgestove.create_cyber_goggles.content.config.CyberConfig;
 import com.simibubi.create.compat.jei.category.SequencedAssemblyCategory;
 import com.simibubi.create.content.processing.sequenced.SequencedAssemblyRecipe;
 import com.simibubi.create.foundation.gui.AllGuiTextures;
@@ -18,18 +18,14 @@ import org.spongepowered.asm.mixin.injection.callback.*;
 public abstract class SequencedAssemblyCategoryMixin {
 	@Inject(method = "setRecipe*", at = @At("TAIL"), remap = false)
 	private void setRecipe(IRecipeLayoutBuilder builder, SequencedAssemblyRecipe recipe, IFocusGroup focuses, CallbackInfo callbackInfo) {
-		if (!CreateCyberGoggles.config.jei.nonrandomScrap) return;
+		if (!CyberConfig.get().jei.nonrandomScrap) return;
 		var size = 8;
 		for (var i = 1; i < recipe.resultPool.size(); i++) {
 			var out = recipe.resultPool.get(i);
 			builder.addSlot(RecipeIngredientRole.OUTPUT, (i - 1) % size * 19 + 15, (i - 1) / size * 19 + 120).setBackground(
 				new IDrawable() {
-					public int getWidth() {
-						return AllGuiTextures.JEI_CHANCE_SLOT.getWidth();
-					}
-					public int getHeight() {
-						return AllGuiTextures.JEI_CHANCE_SLOT.getHeight();
-					}
+					public int getWidth() {return AllGuiTextures.JEI_CHANCE_SLOT.getWidth();}
+					public int getHeight() {return AllGuiTextures.JEI_CHANCE_SLOT.getHeight();}
 					public void draw(@NotNull GuiGraphics guiGraphics, int xOffset, int yOffset) {
 						AllGuiTextures.JEI_CHANCE_SLOT.render(guiGraphics, xOffset, yOffset);
 					}
@@ -45,7 +41,7 @@ public abstract class SequencedAssemblyCategoryMixin {
 	protected abstract MutableComponent chanceComponent(float chance);
 	@Inject(method = "chanceComponent", at = @At("HEAD"), remap = false, cancellable = true)
 	protected void chanceComponent(float chance, CallbackInfoReturnable<MutableComponent> returnable) {
-		if (!CreateCyberGoggles.config.goggles.preciseNumbers) return;
+		if (!CyberConfig.get().goggles.preciseNumbers) return;
 		if (chance * 100 == (int) (chance * 100)) return;
 		returnable.setReturnValue(CreateLang.translateDirect("recipe.processing.chance", chance * 100).withStyle(ChatFormatting.GOLD));
 	}