+++ src/main/java/com/forgestove/create_cyber_goggles/mixin/render/BlueprintOverlayRendererMixin.java
@@ -1,6 +1,6 @@
 package com.forgestove.create_cyber_goggles.mixin.render;
-import com.forgestove.create_cyber_goggles.*;
-import com.forgestove.create_cyber_goggles.content.Util;
+import com.forgestove.create_cyber_goggles.content.config.CyberConfig;
+import com.forgestove.create_cyber_goggles.content.util.Common;
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.simibubi.create.AllItems;
 import com.simibubi.create.content.equipment.blueprint.BlueprintOverlayRenderer;
@@ -42,7 +42,7 @@ public abstract class BlueprintOverlayRendererMixin {
 		int height,
 		CallbackInfo callbackInfo
 	) {
-		if (!CreateCyberGoggles.config.goggles.enhancedStoreRender) return;
+		if (!CyberConfig.get().goggles.enhancedStoreRender) return;
 		callbackInfo.cancel();
 		var mc = Minecraft.getInstance();
 		if (mc.options.hideGui || mc.screen != null) return;
@@ -113,7 +113,7 @@ public abstract class BlueprintOverlayRendererMixin {
 				x += 21;
 			}
 			if (selectedX != 0 && hotbarSelection != null) guiGraphics.blit(hotbarSelection, selectedX - 1, y - 1, 0, 0, 24, 23);
-			Util.renderItemStack(guiGraphics, results.get(index - 1));
+			Common.renderItemStack(guiGraphics, results.get(index - 1));
 		}
 		RenderSystem.disableBlend();
 	}