+++ src/main/java/com/forgestove/create_cyber_goggles/mixin/render/SimpleBackendMixin.java
@@ -1,5 +1,5 @@
 package com.forgestove.create_cyber_goggles.mixin.render;
-import com.forgestove.create_cyber_goggles.CreateCyberGoggles;
+import com.forgestove.create_cyber_goggles.content.config.CyberConfig;
 import dev.engine_room.flywheel.lib.backend.SimpleBackend;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.*;
@@ -8,6 +8,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public abstract class SimpleBackendMixin {
 	@Inject(method = "isSupported", at = @At("HEAD"), remap = false, cancellable = true)
 	private void isSupported(CallbackInfoReturnable<Boolean> returnable) {
-		if (CreateCyberGoggles.config.flywheel.forcedBackend) returnable.setReturnValue(true);
+		if (CyberConfig.get().flywheel.forcedBackend) returnable.setReturnValue(true);
 	}
 }